refactored primitivesToBeHighlighted
